requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
cryptography==3.1.1
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.43.0
django-colorful==1.3
django-cors-headers==3.5.0
django-countries==6.1.3
django-crispy-forms==1.9.2
django-debug-toolbar==3.1.1
django-fontawesome-5==1.0.18
django-host-user-override==0.4.5
django-hosts==4.0
django-markdown-deux==1.0.5
django-pagedown==2.1.3
django-quill-editor==0.1.16
django-subdomains==2.1.0
django-svg-icons==0.1.2
django-upload-validator==1.1.5
djangorestframework==3.12.1
future==0.18.2
gunicorn==20.0.4
huepy==1.2.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==3.1.0
iniconfig==1.1.1
instabot==0.117.0
markdown2==2.3.10
mock==4.0.2
moviepy==1.0.3
numpy==1.19.4
oauthlib==3.1.0
packaging==20.5
paypal==1.2.5
paystackapi==2.0.0
Pillow==8.0.1
pluggy==0.13.1
proglog==0.1.9
psycopg2-binary==2.8.6
py==1.9.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.1.2
python-decouple==3.3
python-magic==0.4.18
python-magic-bin==0.4.14
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
responses==0.12.1
schedule==0.6.0
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
text-validator==0.3
toml==0.10.2
tqdm==4.54.0
tweepy==3.9.0
urllib3==1.25.11
zipp==3.4.0
